+++ test/gurka/test_ferry_connections.cc
@@ -119,11 +119,14 @@ TEST(Standalone, ShortFerry) {
 }
 
 TEST(Standalone, TruckFerryDuration) {
-  // corresponds to 33.3 m/sec (120 km/h) in the below map
-  uint32_t ferry_secs = 9;
+  // corresponds to 21.4 m/sec (~77 km/h) in the below map
+  uint32_t ferry_secs = 14;
 
   const std::string ascii_map = R"(
           A--B--C--D
+          |        |
+          |        |
+          |        |
           |        |
           E--------F
     )";
@@ -151,8 +154,8 @@ TEST(Standalone, TruckFerryDuration) {
   // verify we took the ferry edge and the duration tag was respected
   auto ferry_edge = fastest.trip().routes(0).legs(0).node(1).edge();
   ASSERT_EQ(ferry_edge.use(), valhalla::TripLeg_Use::TripLeg_Use_kFerryUse);
-  ASSERT_NEAR(ferry_edge.speed(), ferry_edge.length_km() / (ferry_secs * kHourPerSec), 0.1);
-  ASSERT_NEAR(fastest.directions().routes(0).legs(0).summary().time(), 45, 0.001);
+  ASSERT_NEAR(ferry_edge.speed(), ferry_edge.length_km() / (ferry_secs * kHourPerSec), 0.2);
+  ASSERT_NEAR(fastest.directions().routes(0).legs(0).summary().time(), 50, 0.1);
 
   // pass a higher ferry cost and make sure it's added to the time
   valhalla::Api higher_ferry_cost =
@@ -160,7 +163,7 @@ TEST(Standalone, TruckFerryDuration) {
                        {{"/costing_options/truck/use_ferry", "1"},
                         {"/costing_options/truck/ferry_cost", "10"}});
 
-  ASSERT_NEAR(higher_ferry_cost.directions().routes(0).legs(0).summary().time(), 55, 0.001);
+  ASSERT_NEAR(higher_ferry_cost.directions().routes(0).legs(0).summary().time(), 60, 0.1);
 }
 
 TEST_F(FerryTest, DoNotReclassifyFerryConnection) {